+++ woninet/core/collectors.py
@@ -151,6 +151,7 @@ def detect_host(
         response = 0
 
     status.latency = response.avg_rtt
+    status.jitter = response.jitter
     status.packet_loss = response.packet_loss * 100
     latency: float = status.latency
     if latency == 0:
@@ -253,6 +254,7 @@ class PingCollector(BaseCollector):
                 dev.mac = status.mac
             dev.latency = status.latency
             dev.packet_loss = status.packet_loss
+            dev.jitter = status.jitter
 
             if dev.reachable:
                 # Only consider reachable hosts as recently seen
@@ -279,6 +281,7 @@ class PingCollector(BaseCollector):
             recorded_metrics = [
                 MetricRecord(ip, "latency", dev.latency),
                 MetricRecord(ip, "packet_loss", dev.packet_loss),
+                MetricRecord(ip, "jitter", dev.jitter),
             ]
 
             if dev.reachable or is_known:
@@ -307,15 +310,20 @@ class PingCollector(BaseCollector):
                     core_logger.error(f"Error in PingCollector for {ip}: {e}")
                     continue
                 else:
-                    latency_metric, packet_loss_metric = future_metric
+                    latency_metric, packet_loss_metric, jitter_metric = future_metric
 
                     latency = (
                         latency_metric.value if latency_metric.value != 0 else None
                     )
-                    packet_loss_metric = (
+                    packet_loss_metric.value = (
                         packet_loss_metric.value
                         if (future_device is not None and latency is not None)
-                        else None
+                        else 0.0
+                    )
+                    jitter_metric.value = (
+                        jitter_metric.value
+                        if (future_device is not None and latency is not None)
+                        else 0.0
                     )
 
                 finally:
@@ -323,4 +331,5 @@ class PingCollector(BaseCollector):
                         device=future_device,
                         latency=latency_metric,
                         packet_loss=packet_loss_metric,
+                        jitter=jitter_metric,
                     )